Add setting to disable pyparsing "packrat" grammar cache
In profiling a significant amount of time is spent updating and querying
a cache data structure used in an attempt to save time parsing. However since
this cache is only used while parsing a single expression and uses locking
for all reads and writes, it had the opposite effect in many cases (actually
slowing down requests).

See #1647